Convert between bibliography formats

Read a bibliography file in one of the supported formats, convert it to annother format and write it to a file.

bibConvert(infile, outfile, informat, outformat)

Arguments

infile
input file, a character string.

outfile
output file, a character string.

informat
input format, see Details.

outformat
output format, see Details.

Details

It should be remembered that there may be loss of information when converting from one format to another.

Bibutils knows two formats for bib files: BibTeX and BibLaTeX and hence can be used to convert between them.

Arguments informat and outformat can usually be omitted, since bibConvert infers them from the extensions of the names of the input and output files.

There is ambiguity for the extension "bib", since it is used for Bibtex and Biblatex entries. For this extension, if informat is missing it is set to "biblatex" but if outformat is missing it is set to "bibtex". This ensures maximum portability, since on input Biblatex is largely a superset of Bibtex. For output the situation is reversed.

For complete list of formats supported by the package, see the documentation of the original bibutils library.

bibentry

Value

The function is used for the side effect of creating a file in the requested format.

References

Chris Putnam, Library bibutils, https://sourceforge.net/projects/bibutils/.

Examples

fn_biblatex <- system.file("inst", "bib", "ex0.biblatex",  package = "bibutils")
fn_biblatex

#> [1] ""

## file.show(fn_biblatex)

## convert a biblatex file to xml
modl <- tempfile(fileext = ".xml")
bibConvert(infile = fn_biblatex, outfile = modl, informat = "biblatex", outformat = "xml")

#> [1] "/tmp/RtmpaXVBwy/filefaac70421c1.xml"

## file.show(modl)

## convert a biblatex file to bibtex
bib <- tempfile(fileext = ".bib")
bibConvert(infile = fn_biblatex, outfile = bib, informat = "biblatex", outformat = "bib")
## file.show(bib)

unlink(modl)
unlink(bib)